moved converters to init so that the import for packages who use us will be easier.
add an option to get a json string
